update UI แจ้งความประชาชน

+++ src/app/layout/layout-desktop/components/issue-online-container/issue-online-check/issue-online-check.component.ts
@@ -3,6 +3,8 @@ import { BankInfoService } from 'src/app/services/bank-info.service';
 import Swal from 'sweetalert2';
 import { DatePipe } from '@angular/common';
 import { IssueOnlineContainerComponent } from '../issue-online-container.component';
+import { BpmProcinstService } from 'src/app/services/bpm-procinst.service';
+import { OnlineCaseService } from 'src/app/services/online-case.service';
 
 @Component({
     selector: 'app-issue-online-check',
@@ -14,12 +16,14 @@ export class IssueOnlineCheckComponent implements OnInit {
     _isShow: boolean = false;
     _isShow2: boolean = false;
     showBank: boolean = false;
+    checknullBankref: boolean = false;
     blockSave = true;
     checkcase : boolean = false;
     popupConsentVisible: boolean = false;
     popupType = 'add';
     popupIndex = 0;
     maxSizeBuffer = 0;
+    formType = "add";
     limitCaseChanelSize = 0;
     maxDateValue:Date = new Date();
     now: any;
@@ -36,21 +40,27 @@ export class IssueOnlineCheckComponent implements OnInit {
         { id: 2, text: "ยังไม่ได้ติดต่อธนาคาร" },
         { id: 1, text: "ติดต่อธนาคาร และได้รับเลขอ้างอิง (Bank Case ID) แล้ว" },
     ];
+    showMoneyWay = false;
 
     constructor(
         private _bankInfoService: BankInfoService,
         private datePipe: DatePipe,
+        private _BpmProcinstService: BpmProcinstService,
+        private _OnlineCaseService: OnlineCaseService
         ) { }
 
     ngOnInit(): void {
         this.maxDateValue.setHours(this.maxDateValue.getHours() + 1);
-        this.setDefaultData();
+        setTimeout(()=>{
+            this.setDefaultData();
+        },200);
     }
 
     async setDefaultData() {
         if (this.mainConponent.formType === "add") {
-            localStorage.setItem("form-index","1");
             this.formData = {};
+            this.formType = "add";
+            this.showMoneyWay = true;
             if(localStorage.getItem("form-blessing")){
                 this.formData = JSON.parse(localStorage.getItem("form-blessing"));
                 this.submission.ways = this.formData.WAY ?? '';
@@ -61,7 +71,29 @@ export class IssueOnlineCheckComponent implements OnInit {
                         this._isShow = true;
                     }
                 }
+            }
+        }else{
+            const _inst_id = Number(localStorage.getItem("inst_id"));
+            const procinstdata = await this._BpmProcinstService.getByInstId(_inst_id).toPromise();
+            sessionStorage.setItem("case_id",procinstdata.DATA_ID);
+            const _case_id = Number(sessionStorage.getItem("case_id"));
+            const bankRef = await this._OnlineCaseService.getBankRef(_case_id).toPromise();
+            this.showMoneyWay = false;
+            this.formType = "edit";
+            this.checkcase = true;
+            this._dataSourcebankref = bankRef;
+            if(this._dataSourcebankref){
+                if(this._dataSourcebankref.length > 0){
+                    this.checknullBankref = false;
+                    this.showBank = true;
+                    this._isShow = true;
+                }else{
+                    this.checknullBankref = true;
+                    this.mainConponent.NextIndex(this.mainConponent.indexTab + 1);
+                }
             }else{
+                this.checknullBankref = true;
+                this.mainConponent.NextIndex(this.mainConponent.indexTab + 1);
             }
         }
     }
@@ -238,22 +270,28 @@ export class IssueOnlineCheckComponent implements OnInit {
     }
 
     SubmitForm(e) {
-        if(this._dataSourcebankref.length <= 0 && this._isShow){
-            Swal.fire({
-                title: "ผิดพลาด!",
-                text: "กรุณาเพิ่มเลขอ้างอิงธนาคาร",
-                icon: "warning",
-                confirmButtonText: "Ok",
-            }).then(() => {});
-            return;
-        } else {
-            this.formData.MoneyWAY = this.submission.moneyWay;
-            this.formData.BLESSING_STATUS = "Y";
-            this.formData.CHECK_BLESSING = true;
-            this.formData.WAY = this.submission.ways;
-            this.formData.BANK_REF = this._dataSourcebankref;
-            console.log(this.formData);
-            localStorage.setItem("form-blessing",JSON.stringify(this.formData));
+        if(this.mainConponent.formType === "add"){
+            if(this._dataSourcebankref.length <= 0 && this._isShow){
+                Swal.fire({
+                    title: "ผิดพลาด!",
+                    text: "กรุณาเพิ่มเลขอ้างอิงธนาคาร",
+                    icon: "warning",
+                    confirmButtonText: "Ok",
+                }).then(() => {});
+                return;
+            } else {
+                this.formData.MoneyWAY = this.submission.moneyWay;
+                this.formData.BLESSING_STATUS = "Y";
+                this.formData.CHECK_BLESSING = true;
+                this.formData.WAY = this.submission.ways;
+                this.formData.BANK_REF = this._dataSourcebankref;
+                console.log(this.formData);
+                localStorage.setItem("form-blessing",JSON.stringify(this.formData));
+                if(e != 'tab'){
+                    this.mainConponent.NextIndex(this.mainConponent.indexTab + 1);
+                }
+            }
+        }else{
             if(e != 'tab'){
                 this.mainConponent.NextIndex(this.mainConponent.indexTab + 1);
             }
